% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rougemont}
\alias{rougemont}
\title{The texts of Denis de Rougemont's 38 books}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 47825 rows and 6 columns.
}
\source{
\url{https://www.unige.ch/rougemont/livres/}
}
\usage{
rougemont
}
\description{
A dataset containing the texts of Denis de Rougemont's digital 
books in a Tidy format.
}
\keyword{datasets}
