% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rougemont_metadata}
\alias{rougemont_metadata}
\title{Metadata of the Rougemont 2.0 project}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2646 rows and 7 columns.
}
\source{
\url{https://www.unige.ch/rougemont/livres/}
}
\usage{
rougemont_metadata
}
\description{
A dataset containing the metadata of Denis de Rougemont's Books
from the list of digital books from the Rougemont 2.0 project 
of the University of Geneva.
}
\keyword{datasets}
